+++ ndk/src/events/content-tagger.ts
@@ -3,45 +3,92 @@ import type { EventPointer, ProfilePointer } from "nostr-tools/lib/nip19";
 
 import type { NDKTag } from "./index.js";
 
-export function generateContentTags(
+export type ContentTag = {
+    tags: NDKTag[];
+    content: string;
+};
+
+export async function generateContentTags(
     content: string,
     tags: NDKTag[] = []
-): { content: string; tags: NDKTag[] } {
-    const tagRegex = /(@|nostr:)(npub|nprofile|note|nevent)[a-zA-Z0-9]+/g;
+): Promise<ContentTag> {
+    const tagRegex = /(@|nostr:)(npub|nprofile|note|nevent|naddr)[a-zA-Z0-9]+/g;
     const hashtagRegex = /#(\w+)/g;
+    let promises: Promise<void>[] = [];
+
+    const addTagIfNew = (t: NDKTag) => {
+        if (!tags.find((t2) => t2[0] === t[0] && t2[1] === t[1])) {
+            tags.push(t);
+        }
+    }
 
     content = content.replace(tagRegex, (tag) => {
         try {
             const entity = tag.split(/(@|nostr:)/)[2];
             const { type, data } = nip19.decode(entity);
-            let t: NDKTag;
+            let t: NDKTag | undefined;
 
             switch (type) {
                 case "npub":
                     t = ["p", data as string];
                     break;
+
                 case "nprofile":
                     t = ["p", (data as ProfilePointer).pubkey as string];
                     break;
-                case "nevent":
-                    t = ["e", (data as EventPointer).id as string];
-                    break;
+
                 case "note":
-                    t = ["e", data as string];
+                    promises.push(new Promise(async (resolve) => {
+                        addTagIfNew(["e", data, await maybeGetEventRelayUrl(entity), "mention"]);
+                        resolve();
+                    }));
+                    break;
+
+                case "nevent":
+                    promises.push(new Promise(async (resolve) => {
+                        let { id, relays } = data as EventPointer;
+
+                        // If the nevent doesn't have a relay specified, try to get one
+                        if (!relays || relays.length === 0) {
+                            relays = [
+                                await maybeGetEventRelayUrl(entity)
+                            ];
+                        }
+
+                        addTagIfNew(["e", id, relays[0], "mention"]);
+                        resolve();
+                    }));
+                    break;
+
+                case "naddr":
+                    promises.push(new Promise(async (resolve) => {
+                        const id = [data.kind, data.pubkey, data.identifier].join(":");
+                        let relays = data.relays ?? [];
+
+                        // If the naddr doesn't have a relay specified, try to get one
+                        if (relays.length === 0) {
+                            relays = [
+                                await maybeGetEventRelayUrl(entity)
+                            ];
+                        }
+
+                        addTagIfNew(["a", id, relays[0], "mention"]);
+                        resolve();
+                    }));
                     break;
                 default:
                     return tag;
             }
 
-            if (!tags.find((t2) => t2[0] === t[0] && t2[1] === t[1])) {
-                tags.push(t);
-            }
+            if (t) addTagIfNew(t);
 
             return `nostr:${entity}`;
         } catch (error) {
             return tag;
         }
     });
+
+    await Promise.all(promises);
 
     content = content.replace(hashtagRegex, (tag, word) => {
         const t: NDKTag = ["t", word];
@@ -52,4 +99,15 @@ export function generateContentTags(
     });
 
     return { content, tags };
+}
+
+/**
+ * Get the event from the cache, if there is one, so we can get the relay.
+ * @param nip19Id
+ * @returns Relay URL or an empty string
+ */
+async function maybeGetEventRelayUrl(nip19Id: string): Promise<string> {
+    /* TODO */
+
+    return "";
 }