+++ ndk-svelte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nostr-dev-kit/ndk-svelte
 
+## 2.0.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nostr-dev-kit/ndk@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes